requirements.txt
pandas
python-dotenv
langchain
langchain-google-genai